@@ -1,4 +1,4 @@
-name: file_picker
+name: pfile_picker
 description: A package that allows you to use a native file explorer to pick single or multiple absolute file paths, with extension filtering support.
 homepage: https://github.com/miguelpruivo/plugins_flutter_file_picker
 version: 2.1.4
@@ -8,13 +8,13 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  pfile: ^0.4.4+3
-  pedantic: ^1.9.2
-  flutter_plugin_android_lifecycle: ^1.0.6
-  plugin_platform_interface: ^1.0.1 
+  pfile: ^1.0.1+2
+  pedantic: ^1.11.0
+  flutter_plugin_android_lifecycle: ^2.0.0
+  plugin_platform_interface: ^2.0.0 
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0 <2.0.0"
 
 flutter:
