@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:file_picker/file_picker.dart';
+import 'file_picker.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'file_picker_result.dart';
@@ -19,8 +19,6 @@ enum FilePickerStatus {
   done,
 }
 
-
-
 /// The interface that implementations of file_picker must implement.
 ///
 /// Platform implementations should extend this class rather than implement it as `file_picker`
@@ -33,7 +31,7 @@ abstract class FilePicker extends PlatformInterface {
 
   static final Object _token = Object();
 
-  static FilePicker _instance;
+  static late FilePicker _instance;
 
   static FilePicker get platform => _instance;
 
@@ -61,14 +59,14 @@ abstract class FilePicker extends PlatformInterface {
   /// with useful information regarding the picked [List<PlatformFile>].
   ///
   /// Returns [null] if aborted.
-  Future<FilePickerResult> pickFiles({
+  Future<FilePickerResult?> pickFiles({
     FileType type = FileType.any,
-    List<String> allowedExtensions,
-    Function(FilePickerStatus) onFileLoading,
-    bool allowCompression,
+    List<String>? allowedExtensions,
+    Function(FilePickerStatus)? onFileLoading,
+    bool? allowCompression,
     bool allowMultiple = false,
-    bool withData,
-    bool withReadStream,
+    bool? withData,
+    bool? withReadStream,
   }) async =>
       throw UnimplementedError('pickFiles() has not been implemented.');
 
@@ -79,13 +77,13 @@ abstract class FilePicker extends PlatformInterface {
   /// of it whenever needed. However, this will force the cleanup if you want to manage those on your own.
   ///
   /// Returns [true] if the files were removed with success, [false] otherwise.
-  Future<bool> clearTemporaryFiles() async => throw UnimplementedError(
+  Future<bool?> clearTemporaryFiles() async => throw UnimplementedError(
       'clearTemporaryFiles() has not been implemented.');
 
   /// Selects a directory and returns its absolute path.
   ///
   /// On Android, this requires to be running on SDK 21 or above, else won't work.
   /// Returns [null] if folder path couldn't be resolved.
-  Future<String> getDirectoryPath() async =>
+  Future<String?> getDirectoryPath() async =>
       throw UnimplementedError('getDirectoryPath() has not been implemented.');
 }
