@@ -4,7 +4,7 @@
 
 // ignore_for_file: lines_longer_than_80_chars
 
-import 'package:file_picker/src/file_picker_web.dart';
+import 'package:pfile_picker/src/file_picker_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
