@@ -19,11 +19,11 @@ class FilePickerResult {
   /// original files (which can be accessed through its URI property).
   ///
   /// Only available on IO. Throws `UnsupportedError` on Web.
-  List<String> get paths => files
+  List<String?> get paths => files
       .map((file) =>
           kIsWeb ? throw UnsupportedError('Unsupported on Web') : file.path)
       .toList();
 
   /// A `List<String>` containing all names from picked files with its extensions.
-  List<String> get names => files.map((file) => file.name).toList();
+  List<String?> get names => files.map((file) => file.name).toList();
 }
